+++ farben/src/core.rs
@@ -1,5 +1,8 @@
-//! Internals that the user may use for certain functions.
+//! Core Farben internals that are used by certain macros.
 //!
-//! This module does not expose anything, just items that the developer may use.
+//! This module is exposed by `farben::*`, which is why that is discouraged. Use `farben::prelude::*` instead to avoid
+//! exposing this module.
 
 pub use farben_core::ansi::style_to_ansi;
+pub use farben_core::debug::tokens_to_markup;
+pub use farben_core::lexer::tokenize;